Replace Javassist dependency with Byte Buddy (issue #6)
The Byte Buddy library is more efficient than Javassist regarding CPU and memory usage. It's also a lot easier to use. This solution doesn't require a custom class loader, which is another advantage.

This commit doesn't include changes to replace the Javassist library when running the application into servlet environments.