Update ByteBuddy and Guice to the latest version
Since ByteBuddy 1.8.4, the default class loading strategy has changed from injection to wrapper. We replaced Javassist with ByteBuddy to avoid the creation of a custom class loader. For that reason, we're now forcing ByteBuddy to inject the changes to the `_NSUtilities`  class directly into the system class loader.

Now WOInject works with Java 8 and 11.